requirements.txt
discord_py_interactions==4.2.1
python-dotenv==0.20.0
